Fixing populate_admin for initial import
The populate_admin script is crashing during initial import due to
unexpected behavior that throws a mismatch of geometry types. This might
be happening because some data derivation between geometry intersections
that result on some geometry of type point instead of linestring. It
doesn't happen on production due to a outdated water_polygon table.

Bug: T205462